requirements.txt
git+https://github.com/mitsuhiko/flask.git
git+https://github.com/xbrowse/xbrowse.git
Flask-Script==2.0.5
ipython==2.1.0
pymongo==2.7.1
